➕ Add: Base, CodeBaseCVC 파일 추가 (#38)
